requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
PyPDF2==3.0.1
python-docx==1.1.0
plotly==5.18.0
seaborn==0.13.0
matplotlib==3.8.2
wordcloud==1.9.3
regex==2023.12.25
sentence-transformers==2.2.2
scipy==1.11.4
Pillow==10.2.0